Updated to rsyslog 8, sourced from upstream project's yum & apt repos
Tracker story: https://www.pivotaltracker.com/story/show/93045744

+++ stemcell_builder/stages/base_centos_packages/apply.sh
@@ -13,7 +13,7 @@ case "${stemcell_operating_system_version}" in
     ;;
   "7")
     init_package_name="systemd"
-    version_specific_packages="nmap-ncat rsyslog rsyslog-relp rsyslog-gnutls rsyslog-mmjsonparse"
+    version_specific_packages="nmap-ncat"
     ;;
   *)
     echo "Unknown centos version: ${stemcell_operating_system_version}"
@@ -40,17 +40,19 @@ cmake sudo libuuid-devel parted NetworkManager e2fsprogs"
 pkg_mgr install ${packages} ${version_specific_packages}
 
 # Install runit
-pkg_mgr install "rpm-build rpmdevtools glibc-static"
-cookbook_release=1.2.0
 runit_version=runit-2.1.1
-run_in_chroot $chroot "
-  curl -L https://github.com/opscode-cookbooks/runit/archive/v${cookbook_release}.tar.gz > /tmp/v${cookbook_release}.tar.gz
-  tar -C /tmp -xvf /tmp/v${cookbook_release}.tar.gz
-  tar -C /tmp -xvf /tmp/runit-${cookbook_release}/files/default/${runit_version}.tar.gz
-  cd /tmp/${runit_version}
-  ./build.sh
-  rpm -i /rpmbuild/RPMS/${runit_version}.rpm
-"
+if ! pkg_exists ${runit_version}; then
+  pkg_mgr install "rpm-build rpmdevtools glibc-static"
+  cookbook_release=1.2.0
+  run_in_chroot $chroot "
+    curl -L https://github.com/opscode-cookbooks/runit/archive/v${cookbook_release}.tar.gz > /tmp/v${cookbook_release}.tar.gz
+    tar -C /tmp -xvf /tmp/v${cookbook_release}.tar.gz
+    tar -C /tmp -xvf /tmp/runit-${cookbook_release}/files/default/${runit_version}.tar.gz
+    cd /tmp/${runit_version}
+    ./build.sh
+    rpm -i /rpmbuild/RPMS/${runit_version}.rpm
+  "
+fi
 
 # arrange for runit to start when the system boots
 if [ "${init_package_name}" == "systemd" ]; then
@@ -58,6 +60,11 @@ if [ "${init_package_name}" == "systemd" ]; then
   run_in_chroot ${chroot} "systemctl enable runit"
   run_in_chroot ${chroot} "systemctl enable NetworkManager"
 fi
+
+# we need newer rsyslog; this comes from the upstream project's own repo
+echo ${stemcell_operating_system_version} >${chroot}/etc/yum/vars/releasevernum
+cp $(dirname $0)/assets/rsyslog.repo ${chroot}/etc/yum.repos.d/
+pkg_mgr install "rsyslog rsyslog-relp rsyslog-mmjsonparse rsyslog-gnutls"
 
 exclusions="mlocate firewalld"
 pkg_mgr erase $exclusions
